* Removed the option to use M$ for pre-processing. (It wasn't implemented right anyway and we rely on gcc for everything else.)  So the option -gcc and -vc no longer exist.
* Added -D as a command-line flag -- the defines just get passed to the
pre-processor.

* Added some usefule cleaning targets to Makefiles -- lib, src, and tests
now have a "clean_prefix" that only removes generated files with the
OUT_PREFIX prefix.  And the top Makefile has a "clean_src" and
"clean_src_prefix" that just do the obvious thing in lib and src.

I tried to figure out why the column information for errors is often
wrong, but no luck yet.

By the way, the huge line counts for changes in bin/genfiles is because
a change to the Cyclone that adds or removes a C label or temporary
variable causes all the numbers to shift for the rest of the file.  C'est
la vie.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@798 581db07f-6e62-4d40-9733-d432c4eebb07