Introducing a sound system for taking the address of local variables. It's pretty minimal still, and the concrete syntax and error messages leave much to be desired.  Nonetheless, I intend to take out the "address of local" warning real soon since it should be safe.  Next on the to-do list is a similar trick for reference patterns.  I've implemented but not tested things like &x.1 and &y[3] where x and y are on the stack.
One restriction of note: in C, &e1[e2] is always allowed.  When e1 is
a FixedArray or a Tuple (not a pointer to a tuple), I am requiring e1
to be an lvalue.  I doubt anyone will mind, since we are only talking
about when it appears under &.

Here is the "hello world" of stack allocation:

void swap<`a,`r1,`r2>(`a @`r1 x, `a @`r2 y) {
  `a temp = *x;
  *x = *y;
  *y = temp;
}

int main() {
  int z = 4;
  int q = 2;
  swap(&z,&q);
  printf(" %d %d ", z, q);
  return 0;
}

Notice that region variables are type variables of region kind.  In a
function prototype, the compiler infers the kind of a type variable to
have region kind if it appears in an appropriate position (juxtaposed with
* or @).  All other type variables are inferred to have kind BoxUKind.
A pointer without a region annotation means a pointer into the heap.

There is almost no other concrete syntax at this time, but thanks to
type inference, we can do some more interesting things.  This file
generates just the right errors:

int g(int @ a) {
  return *a;
}
int h<`r>(int @ `r a) {
  int @ `r @ x = &a; // no
  _ y = &a; // okay -- we infer that y has type int @ `r @0
  g(a); // no
  return *a + **y; // okay
}

struct Ptr<`a> { `a @ v; };

int glbl = 7;

void f() {
  int x = 3;
  int @ y = &x; // no
  _ z = &x;     // okay
  int * `r q = &x; // no
  g(z);  // no
  h(z);  // okay
  g(&x); // no
  h(&x); // okay
  Ptr(z); // no
  g(&glbl);
}

Stack regions are numbered which is why strange numbers show up in
error messages.  Integrating some concrete syntax shouldn't be too
bad.  Adjacent declarations (eg, x, y, z, q in function f above)
belong to the same region.

Other changes:

* In order to get cyctest to go through, I had to change the types of
String::to_int and String::to_double so that they could accept stack
variables.

* Changed cyc_main in cyc_test to return an int so that the Makefile
  doesn't fail due to a bad exit code.

* Lots of unrelated housekeeping things.  If you're looking for the
  stack-allocation stuff, a good place to start is Tcexp::tcAddress,
  although a lot of the magic is integrated into unification.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@882 581db07f-6e62-4d40-9733-d432c4eebb07