First, some little things:
* I defined Position::seg_t and Tcenv::tenv_t to be typedefs that are
  identical to Position::segment and Tcenv::tenv respectively.  I
  prefer the former because of emacs highlighting, but I resisted
  making such a pervasive change.  But then I realized typedefs were
  transparent, so I could have both and migrate code at my leisure,
  i.e. when I'm looking at it for some other reason anyway.

* Created Tcutil::resolve_struct_designators.  This code is fairly
  heinous.  But it was essentially being duplicated to type-check
  Struct_e (the expression) and Struct_p (the pattern).  So the
  heinous-reduction factor was 50%. :-)

Now the good news:

I believe that I now type-check patterns soundly with respect to stack
allocation, so we no longer warn about reference patterns.  But this
is all very subtle, so please keep an eye on this code.  Note that
code of the form "let *x = e" and an analogous switch will be rejected
because e might not be an lvalue.  It should be straightforward to
check if e is an lvalue and act accordingly.

A little primer on how patterns are now type-checked: tcPat takes an
addition parameter which is either null or a pointer to the region on
which we're pattern-matching (sorta).  Null means reference patterns
are not allowed and we post an error.  If not null, the Reference_p
case gives its variable a pointer type with the region that was passed
to tcPat.  When descending under Tuple_p and Struct_p we pass along
the same region_opt.  When descending under Enum_p and Xenum_p we pass
along the heap region because all Enums and Xenums point into the
heap.

The interesting case is Pointer_p.  We don't yet know what region the
underlying pattern should be in because we haven't unified the pattern
against an expression yet.  So the region we pass along is a fresh
wildtyp() and we put the same region in the PointerType we build for
the outer pattern.

Now a Pointer_p means a dereference, so we must make sure the region
is accessible.  So after tcPat and the ensuing unification in tcStmt, we
call the new Tcpat::check_pat_regions which just crawls over the pattern and
calls check_rgn_accessible on all those wildtyp() regions we created
on the previous pass.  (Where are they? They're in the second field of
the PointerType that is in p->topt->v for a Pointer_p.)

Like I said, things are a bit subtle.

Here's an example that gets the right error message:

int main() {
  int *_ x = null;
  if(1) {
    int y = 17;
    x = &y;
  }
  switch (x) {
  case null: return 0;
  case &z:   return 1; // not safe!
  }
}

Suppose the last pattern were &*z instead.  You might be tempted to
say this is sound, although any dereference of z would not be.  But
the way we translate patterns would actually dereference x, so it's
correct that we reject this strange case.

KNOWN UNSOUNDNESS: This actually has nothing to do with patterns -- if
a function f has a region variable `p, we should not allow `p to be
instantiated with an inaccessible region.  We need to check at each
call site that the regions mentioned in the argument types are still
accessible.  Here's code that wrongly passes:

void f<`p>(int *`p x) { ; }

int main() {
  int *_ x = null;
  if(1) {
    int y = 17;
    x = &y;
  }
  f(x);
  return 0;
}

I don't think I'll have time to fix this.  Greg, you're welcome to
take a stab at it. :-)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@888 581db07f-6e62-4d40-9733-d432c4eebb07