Lots of fixes with kinds and kind inference:
* kinds now include:
  A -- any type
  M -- "mem" types, same as A but excludes abstract structs
  B -- "box" types, types from M that go in a gp register
  R -- regions
  E -- effects
with the obvious subkinding (B <= M <= A).  UnresolvedKind has
gone away.  Instead, we use a conref<kind_t> to keep track of
the kinds of type variables during kind inference.  So,
UnresolvedKind is represented as a No_constr.  We needed the
conref's to get the right sharing behavior for kind inference.

* abstract structs and type variables of kind A can only occur
under pointers.  Well-formedness now checks types w.r.t. a kind
to achieve this.  This triggered a bunch of other hidden bugs.

* for structs and enums, you have to put the kinds on non-B bound
type variables.  For instance, you can't just write:

  struct List<`a,`r> {`a hd; struct List<`a,`r> * `r tl;};

but rather, you have to write:

  struct List<`a,`r::R> {`a hd; struct List<`a,`r> * `r tl;};

This isn't so bad anyway since if you make the struct abstract,
you have to put the kinds in.

For typedefs, you don't need to write kinds on the bound
type variables.  The issue is one of recursion.  It's possible
to do the kind inference without the explicit kinds, but would
require me tracking both upper and lower bounds on type variables.
Ultimately, I think I'm going to have to do this (see below).

* structs, functions, enums, and the fields in enums and xenums
can abstract type variables of all kinds except M.  For instance,
one can write:

  int ptrcmp<`a::A>(`a *x, `a *y);

but you can't write:

  int ptrcmp<`a::M>(`a *x, `a *y);

On the other hand, typedefs can abstract variables of all kinds,
including M.  So, for instance, you can write some useful typedefs
that abstract things besides boxed types:

  typedef $(`a,`b) @pair<`a::M,`b::M>;

  typedef `a ?array<`a::M>;

and use them just about anywhere (e.g., pair<$(int,int),char[3]> or
array<$(int,int)>).

The reason is that the typedef gets compiled away via substitution
before translation to C.  So, all of the variables of kind M will
disappear.  The translation to C only knows how to translate A and
B kinds, which it represents as an abstract struct and void*
respectively.

If we wanted to give up separate compilation (and polymorphic recursion,
and first-class polymorphism), then we could allow abstraction over
the M kind and implement via compile-time specialization (e.g., as
in C++.)  That is, I can imagine adding support for second-class
templates that abstract M types.  For instance something like this:

  template ListGen<`a:M> {
    struct List{`a hd; struct List *tl;}

    int length(struct List) { ... }
    ...
  }

The understanding is that template specializations would be generated
at compile time.  So, for instance we might have a declaration of the
form:

  namespace ListIntPair = ListGen<$(int,int)>;

In fact, from this, we could generate the boxed List implementation as:

  namespace List<`a::B> = ListGen<`a>;

It's a thought.  I think the way you'd implement this is if you wrote:

  namespace List<`a::B> = extern ListGen<`a>;

then you'd only generate the type declarations and (extern) prototypes
from the template within the client.  If you leave the extern off, then
you'd get the full thing specialized appropriately.  That way, the
programmer has control over where the code is generated (and how many
copies there are.)  In fact, this suggests a fairly nice way to incoprorate
a bunch of other neat type-theoretic ideas, including functors or
compile-time type analysis (a la lambda-ML-i) to get at a more
principled C++.

* Unfortunately, there's still a slight problem in that the compiler
currently allows you to write:

  int ptrcmp(`a::M *x, `a::M *y);

which makes it appear as if you're abstracting an M type.  You're
not.  What happens is that during kind inference, we calculate
upper bounds on the type variables in a bottom-up fashion.  At
the point where we figure out what type variables to abstract,
we drive the upper bounds of M type variables down to B.  So
when you write the above, this is elaborated into something like
this (ignoring the effect):

  int ptrcmp<`a::B>(`a::B *x, `a::B ::*y);

which succeeds instead of generating an error.  The problem is that
the kind inference can't tell the difference between internally
generated upper bounads and explicit ones placed on type variables by the
user.  The solution is to add upper and lower bounds on type
variables and treat user annotations as lower bounds.

* There are some more bugs for me to track down.  In particular, Dan's
stack allocation of the options in cyclone.cyc doesn't work for some
reason that I need to figure out.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@964 581db07f-6e62-4d40-9733-d432c4eebb07