This is a big check-in.  Let me first dispense with the little stuff (so skip ahead to get the good stuff)
* Added Dict::iter2_c, Dict::intersect_c, and Dict::forall_intersect.
  For the first two, my client code exploits that when f is called,
  the first argument of type `b is from the first dict.  Of course,
  that's a pretty dangerous assumption, but it's true.

* Added List::list_cmp.  Had to change the list_cmp in
  tests/cyc_test.cyc to be called test_list_cmp.

* Fndecl now carries a list of vardecls corresponding to the
  parameters.  This is set by the type-checker and used by it and
  later passes.  By having this list, we can use pointer-equality on
  vardecls.  So don't ever make a Vardecl except at a binding
  occurrence.  This was already true except for parameters.

* Fallthru_s now carries the destination switch_clause instead of the
  body of the destination clause.  (My code needed the patterns vars
  of the destination.)

* We now use the new control-flow checker.  If you run into problems,
  do the obvious line-change in cyclone.cyc.  The new checker isn't
  currently preventing goto's into scope for obscure coding reasons.
  In my next check-in, the old control-flow checker will be completely
  bypassed and this check will be in there again.

Okay, now a brief intro to the new control-flow checker:

the new control-flow checker should have the following
"conservative extension" properties:

* It's sound -- you cannot ever read memory that hasn't been written
  to with a value of the correct type.

* It accepts any program the old control-flow checker accepted.

The main new technology is:

* You can take the address of uninitialized memory.

* The checker tracks must-aliases in a reasonable (?) manner.

So here's code that now checks and didn't before:
  int x;
  int *_ y = &x; // would have been rejected right here.
  *y = 3;
  return x;

It works with malloc too:
  struct Foo<`r::R> { int *`r x; }

  int x;
  _ y = new{Foo(&x)};
  x = 7;
  return y;

This will make porting C code tons easier because we will hack the
front-end to support malloc(sizeof(t)) just like we hacked it to
support printf(" ... ", ...).  And the checker already has the
infrastructure to support:
  foo * x = (foo *)malloc(sizeof(foo));
  x->f1 = ...
  ...
  x->fn = ...
After all, internally the checker is treating the "new" in the example
above as a sequence basically equivalent to a malloc followed by field
assignments.  And with a reasonable flow analysis, there aren't ad hoc
restrictions about how quickly the fields get initialized.

That's about all you need to know to hack, but you should
also read the limitations below.

Here is a very brief overview of the architecture.  I don't really
expect you to understand it because it is way too brief.

The analysis makes two passes over a function.  In the first, it takes
each expression and creates an absexp (abstract expression) that
characterizes the expression's assignments and uses.  It decorates
statements with the absexps.  The second pass never looks at exps,
only absexps.

The second pass is an iterative data-flow analysis.  At each
program point we attach a "flowinfo" which characterizes
initialization states and must-aliases.  An absexp is a flowinfo
transformer.  Toss in a join operation for flowinfos and you're done.

So all the interesting stuff is in the answers to five questions:
* What do flowinfo's look like?
* What do absexp's look like?
* How do we interpret absexps as flowinfo transformers?
* How do we take the join of two flowinfo's?
* How do we abstract exps to absexps?

Soundness amounts to getting the "how" questions right.

A flowinfo maps "places" to "initialization states".  In the absence
of (unboxed) structs and (unboxed) tuples, a place is a local variable
or a malloc point (including enum and xenum expressions).  If one of
these "roots" has struct or tuple type, it is not a legal place --
only its "fully expanded paths" are places.  That is, x.f1.f2 is place
if the type of x.f1.f2 is not a struct or tuple.

An initialization state is either "must point to p'" where p' is a place
or "unknown of escapedness cross initlevel".  If p is Escaped, it
means we don't know about all the pointers to p.  Else p is
Unescaped.  There are three initlevels: None means p may be
uninitialized. This means p is initialized but any of the fields of
the thing it points to may be uninitialized.  All means everything
reachable from p is initialized.  Obviously you could add any number
of other initlevels, but this domain seems like a sweet spot.

A simple absexp is one of: skip, bottom , use(x),
assign(x,y), malloc(prog-point).  x and y are abstract operands which
I won't describe fully here.  Basically, they could be Unknown, a
local, a malloc point, or a restricted combination of &, *, and .
applied to an abstract operand.  There are also "grouped" absexps
where the group flavor is one of "ordered", "unordered", or "one of".
Hopefully the meaning of groups is intuitive.

Some rough intuition for the transform: use makes its operand escape.
(If something escapes, it must be fully initialized.)
assign(x,y) transfers y's initialization state to x.  If x is unknown
or escaped then y must be fully initialized.  Otherwise, aliasing
causes unsoundness.

The transform for Unordered is a kludge.  The right thing is to take
the join over all permutations of the group.  Any sound thing must
approximate doing this because there may be exactly one permutation
that is unsound.  One such approximation is to say the elements of the
group might be executed in any order and any number of times
(including 0).  This is sound, but unfortunately it rejects too much
code, such as when a malloc-and-initialize (eg. Foo(7) for enum
variant Foo) appears as a function argument.  So after doing the "any
number of times" thing, we go back through "re-execute" all the
mallocs and assignments to malloc-pt fields.  This is sound because
these operations are all independent of each other so it doesn't
matter what order they happen in.  Did I mention that this is a
kludge? :-)

The join of flowinfos is straightforward except if p must point to p'
in one flowinfo but not in the other.  In such a case, p becomes
unknown and everything it points to in either flowinfo escapes.

The abstraction function for exps was probably the hairiest part.
Fortunately, when it gets too tough, you can soundly take a
subexpression and assume that it is eagerly used (escapes) and
evaluates to an unknown absexp.  This is the source of the more
noticeable limitations.  Here are two examples:

 * If you cast a value of struct or tuple type, the value will escape.
 * All the expressions in e1 ? e2 : e3 get used eagerly.  This is
   clearly conservative:
         int x;
         int *y = &x;
         int *z = b ? y : y
   gets rejected.  I'll be happy to explain why it's a pain to do
   better.

There is kludge in the abstraction function w.r.t. undefined
evaluation order.  Consider
         int x;
         int *y = &x;
         int **z = *y;
         f(z, **z = 17);
This code is safe, but won't be if we think that z might escape before
**z = 17 occurs.  So the abstraction does the escape _after_ the
expressions, so the abstraction looks like:
   ordered
     unordered
       skip // for z
       **z = Unknown // for **z = 17, we don't track int values
     unordered
       use z   // for passing first parameter to God-knows-where
       use **z //    " "      second  "  "       "
But in fact, the second unordered is spit out as an ordered, not
because the parameters get passed in order, but because their order
_doesn't matter_.

Okay, that's enough rambling.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@969 581db07f-6e62-4d40-9733-d432c4eebb07