Added unions, or rather what was easy to add.  You can define unions as long as the members only contain "bits" (i.e., char, short, int, long long, float, double, tuples/structs/unions of bits, enums with no value-carrying constructors.)
There's no direct way to build a union, but you can cast a value
of type t to a union type as long as there's a member of type t
in the union.  Note that in C, the only valid initializer for
unions is the first member's type.  We don't currently check this,
but GCC barfs if you use any other member.

Struct-member and Struct-arrow are now overloaded to support
unions.  We should probably add a flag to these expression
forms to indicate (after type-checking) whether its really
a struct or a union.

Dan, you'll need to fix the control-flow stuff.  Right now,
I get errors if I try to assign a member in a union.  (I just
haven't bothered to figure out the right way to do this and
figure you can do it easier.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@983 581db07f-6e62-4d40-9733-d432c4eebb07