Changed translation to C of topdecls.  Now everybody builds on the same global list of result_decls.  So there is no more add_decls or temp_topdecls.  This was necessary to have a struct type (not a pointer to struct type) as a tuple type field.  In other words, hoisting tuples to the top of the file is wrong.
To test this out, tqual_t is now struct Tqual, not struct Tqual *.
This is inconvenient in that you can't make a list of tqual_t's now,
or call snd, or List::split on a list of $(tqual_t,type_t)@.  It would
be fun at some point to make tqual_t a union of an int and a struct,
but I suppose there are more important things to do. :-)

Removed the shadow field from vardecl_t's.  We weren't using it.

--Dan


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1043 581db07f-6e62-4d40-9733-d432c4eebb07