Finished my changes to TcPat that pass down a topt.  Actually, there's one thing left to do: when you see &Foo(...) for tunion variant Foo and the topt is the tunion (not the tunion field), I'm not passing down a topt right now.  The defaults are such that you'll probably never notice -- what's being forgotten is info about the instantiation of a polymorphic tunion.
Also fixed problem with existential vars: In pattern Foo<`a>, we
weren't substituting VarType(`a) for the `b where `b is the type
variable in the variant's definition.  Nobody noticed because we only
have one unpack in the whole compiler and it (sensibly) uses the same
tyvar as the definition.

WARNING: I haven't modified the translation to C yet to account for
the fact that, thanks to passing down type info, the type for pattern
&Foo might be "pointer to variant Foo", not "tunion Bar".  So the
translation will throw an "impossible" error.  I'll fix it tomorrow.

Some minor things:

* took out an unused field in Tunion_p

* now share EmptyAnnot rather than mallocing a new one for each
statement.  By the way, I think XTUNION VARIANTS SHOULDN'T HAVE TO MALLOC.
Just use their tag-address, but make sure the first word
at the address is 0.  That way value-carrying variants can just follow
the pointer and then fail the ensuing tag-check.

Making this change would further reduce the tunion/xtunion
differences.  We'd be down to changing the exhaustiveness check and
the generation of tags.  In fact, we could translate tunion like the
xtunion I sketch above, but I'd want to check that the performance hit
is small.

* Regionified resolve_struct_designators


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1044 581db07f-6e62-4d40-9733-d432c4eebb07