Preparation for the printf and scanf stuff:
1. fixed bug in treatment of pointer '+'.  Consider the code:

int buf[10] = {...};
int ? _ x = buf+10;
x[-9];

This was giving an array bounds error because we treat "buf+10"
as a pointer to an array of size 0 (i.e., int@{0}) and then convert
it to a tagged array.  Now we convert the buf to a tagged array
and then add 10 to it.

2. made the "silent" coercions a bit more permissible w.r.t. region
lifetimes.  You can cast a pointer or tunion (implicitly) to another
pointer/tunion as long as the lifetime of the original is as long as
the target.

3. made injection a bit more permissible.  We still try for an exact
match first, and failing this, use the same coercions that we use
when calling a function (e.g., arithmetic promitions, silent casts,
etc.)

4. region-ified the math.h library.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1169 581db07f-6e62-4d40-9733-d432c4eebb07