We now support the GCC-style "format" attribute which allows us to statically type-check certain vararg functions against a format descriptor string literal.  In particular, when you call printf, scanf, etc. now and use a string literal for the format descriptor, we check that the actual argument types match up with the descriptor. You can also define your own functions to be type-checked like this.
The attribute has the form:

  __attribute__((format(printf,n,m)))

or

  __attribute__((format(scanf,n,m)))

Use "printf" for printf-like functions and "scanf" for scanf-like
functions.  The parameter n indicates the position (starting from 1)
of the format string argument.  The parameter m indicates the
position of the first argument to type-check against the format
string.  When m is 0, then only the format string is checked for
consistency.  This gets used in the vprintf-like functions where
you pass in an array of arguments instead of a vararg list.

If you call a function with this attribute without a string literal,
then no static-checking is done.  Note however, that the dynamic
checking happens regardless.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1185 581db07f-6e62-4d40-9733-d432c4eebb07