This works under Linux now.  Could someone else test it to sanity-check? I think we only need bash and cat; I checked and these are installed onthe INRIA Linux box.  There are some tunable params folks might want to change:
SLOP       we will "wait" TLIMIT-SLOP secs before cancelling
INTERVAL   we poll every INTERVAL secs to see if we're done (thus
           we could end up waiting TLIMIT-SLOP+INTERVAL-1 secs) by
           checking if the output file size is nonzero.

I tried using sleep + trap with signals, but it was flaky at best, and
with polling we don't have to fork another process.

Finally, I piped the output of validate.exe to /dev/null.  We might
consider putting its output on stderr, or not eliminating it, just
so the INRIA guys can see we have this slick tool. :)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1278 581db07f-6e62-4d40-9733-d432c4eebb07