I added a README file that explains VERY basically what is going on; this is required for submission.  Those of you who did more of the work should feel free to add to it.
We also need to come up with a cool name for our team, and a cool
name for our program; both of these go in the entry form.  If no one
suggests anything better, we will be Team Snowstorm and our program
will be called Acetone.  They're cheesy, I know ...


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1285 581db07f-6e62-4d40-9733-d432c4eebb07