Well, it's all up to Mike now.
If you run make CHECK_STACK=x, we'll generate code that checks in
(hopefully enough) places that we don't personally use more than 7Mb
of stack.  Add in start-up, GC, and incidentals, and we should still
be under 8Mb.

Run make without the definition and you'll get a somewhat faster
optimizer.

I haven't messed with the script or anything, so I leave that to Mike.

Well, we did what we came to do.  I just hope to end up somewhere
respectable in the peleton.  Thanks everybody. :-)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1290 581db07f-6e62-4d40-9733-d432c4eebb07