This is a midway check in for the new approach to effects.  Here's what's changed and some notes on current problems:
(1) regions that are ommitted are replaced with "_".  Note that "_"
is treated quite differently now and is resolved in a context-dependent
fashion (see below).  Thus, leaving off a region no longer defaults
to the heap in all contexts.

(2) typedefs allow you to under-specify the type arguments.  So, for
instance, if list_t takes two arguments, then writing list_t<int>
is short-hand for writing list_t<int,_>.  Similarly, list_t alone
is short-hand for list_t<_,_>.  The "_"'s are resolved in the
context-dependent fashion described below.  As a result, the libraries
no longer have "glist_t" and "gdict_t" typedefs.  We rarely need them
anymore.  (In some situations, it might be convenient to have
hlist_t or hdict_t though, but less so than before.)  I'm likely
to add support for this kind of abbreviation to structs, tunions,
etc.  It's extremely useful.

(3) string_t, mstring_t, stringptr_t, mstringptr_t are all now
region-ified.  So the definitions in core.h now read:

  typedef const char ?`r string_t<`r>;
  typedef char ?`r mstring_t<`r>
  typedef string_t<`r> @`r2 stringptr_t<`r,`r2>;
  typedef mstring_t<`r> @`r2 mstringptr_t<`r,`r2>;

This will allow us to change string.h back to using the nice
abbreviations.  For instance, we can now write:

  extern mstring_t<`r> strcpy(mstring_t<`r> dest, string_t src);

(4) An occurrence of "_" in an argument position of the type of a
function is replaced with a fresh type variable.  Like other free
type variables in the function's prototype, these type variables
are abstracted at the top-level.  So, for instance, if you write:

  void foo(_ x);

this is short-hand for:

  void foo<`a::B>(`a x);

Regions and effects are treated the same way.  So if you write:

  void foo(int *x, int *y);

Then this is treated the same as:

  void foo<`r1::R,`r2::R>(int *`r1 x, int *`r2 y);

Thus, List::length can have the following deceptively simple prototype:

  int length(list_t x);

This is because list_t is expanded to list_t<_,_> which is then
replaced with list_t<`a,`r> where `a and `r are fresh type variables.

(5) If "_" occurs in a result type of a function, then what we do
depends upon the kind: If the "_" has kind RgnKind, then "_" is
replaced with `H.  If the "_" has kind EffKind, then it is replaced
with the empty effect.  And finally, if "_" has kind B, M, or A (a
type), then an error is reported.  In other words, within a function
prototype, "_" is never interpreted as an evar.  This gets around a
number of potential problems and cleans up the well-formedness code
considerably.  However, as a side effect, I had to change bison which
was generating functions with "_" as the return type in order to avoid
figuring out what the type of a production really was.  Instead, I
went back to the old approach of using a macro to extract values and
do the "down-cast".  As a result, the parser takes a lot longer to
compile.  [Dan, since I use a nested statement expression in the
macro, it may cause the VC version of the system to choke...]
When we make Trevor's proposed changes to Bison, this problem will
go away.  In particular, if we just specify:

%type <t> NT1, NT2, ..., NTN

in the grammar where t is an arbitrary type, and forget the whole
tunion thing, then we have enough information to generate a constructor,
and down-cast functin.  This should be simple enough to do but
I had enough on my plate for now.

(6) An occurrence of "_" within the type of a local variable is
replaced with an evar and can then be constrained with unification.
So, for instance, one can write:

void foo(int i) {
  int *p = &i;
  ...
}

(7) A free occurrence of "_" within a top-level variable's type, or
within a type definition (i.e., struct, typedef, tunion, etc.) is
treated in the same fashion as return types above.  That is, region
"_"'s are replaced with the heap, effect "_"'s are replaced with
the empty effect, and "_" of a type kind is considered illegal.
(This plugs a big hole that we had before.)  I'm likely to change
the treatment of region "_"'s within type definitions later on.

(8) Default effects are now calculated in an entirely different
way.  There is a new effect constructor regions(-) which maps
a type to a set of regions.  The default effect for a function
is now regions(res)+regions(arg1)+...+regions(argn)+regions(varargs).
No slack effect variables are used in the default effect.  So,
for instance, the function map which has the following abbreviated
prototype:

  list_t<`b> map(`b f(`a), list_t<`a>)

elaborates to the following:

  list_t<`b,`H> // unspecified region in return type defaults to heap
    map(
     `b f(`a; regions(`a)+regions(`b)), // effect for f is regions of args+res
     list_t<`a,`r>; // unspecified region in arg defaults to fresh tyvar
     {`r}+regions(`a)+regions(`b) // effect includes r & regions of `a & `b
    )

As a result, the ADTs that used to have abstract an effect variable
(set, dict, arg) no longer do, and as a result have a much simpler
and more intuitive interface.

Note however that the compiler still supports effect variables in
case we need them.  And I've left all of the old code for calculating
the default effects in there in case we decide we don't like this
approach.

I have ported all of the libraries, the compiler, bison, and cyclex
to work with the new scheme.  This mostly consisted of (a) replacing
glist_t with list_t, gdict_t with dict_t, etc., (b) occasionally
putting a `H on an argument to a function.  I was sometimes surprised
when `H was not needed.

One problem that came up which I had not anticipated is the following:
Sometimes when you declare a local variable using "let" or "_", the
unification is a bit too eager to constrain the region to something
you don't expect.  For instance, consider:

exp_t array_exp(list_t<exp_t,`H> es, seg_t loc) {
  let dles = null;
  for (; es != null; es = es->tl)
    dles = new List(new $(null,es->hd),dles);
  dles = List::imp_rev(dles);
  return new_exp(new Array_e(dles),loc);
}

This fails to type-check because at some point, the unifier has had to
nail types down before it really should and it guesses the wrong type
for the "null" nested within the tuple.  Consequently, I had to put in
more type information to direct things.  This came up in a few places
and I'll try to track down what the problem is and fix it.

I have not yet ported the tests directory, and the error messages
are now really confusing.  You're likely to get a message like
"string_t != string_t" because the pretty printer doesn't yet know
about the abbreviated typedefs.  Nevertheless, this is a pretty
good stopping point.  Feedback is appreciated.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1294 581db07f-6e62-4d40-9733-d432c4eebb07