Added script to make a distribution, and a file .excludelist that contains the files not to include in the distribution.  You call makedist with a version # as its argument, and it will create the file cyclone-VER.tar.gz in the toplevel directory.
In the future, we might like this script to automatically generate
version #'s, and tag CVS with that version.  We might also like the files
to be untarred in a directory cyclone-VER, rather than just cyclone.
Speaking of which, is there a version # in a .h file somewhere so
we can ask what version the compiler is?  Seems like we want this.

Feel free to add to the .excludelist file; just put in regexp's of
files you want to avoid including.  The patterns exclude files who
have any part of the pattern in their path.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1360 581db07f-6e62-4d40-9733-d432c4eebb07