Added some more comments to absyn.h.  We ought to re-organize this in a top-down fashion.
Added a "guide to porting from C" section to the manual.  There's
some duplication with the FAQ, but I think it's good to have a
list of things that you know you'll have to do.  There's also
some details in here about how to build wrappers and write
interfaces to existing C code.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1504 581db07f-6e62-4d40-9733-d432c4eebb07